pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptdesign"
version = "0.1.0"
description = "A CLI tool with web UI for evaluating prompts and suggesting improvements"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "typer>=0.9.0",
    "anthropic>=0.18.0",
    "pydantic>=2.6.0",
    "markdown-it-py>=3.0.0",
    "textstat>=0.7.3",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.2.0",
]

[project.scripts]
prompteval = "cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["cli", "backend"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
